feat(retry): add `Retry` policy interface (#25)
Add a new interface, `sghi.retry.Retry` and a few of its
implementations. A retry policy allows an application to handle
transient failures such as when trying to connect to a network
resource or another service. This is achieved by transparently
repeating the failed operation until it succeeds or until some
criteria are met.