chore(retry): rename the `exponential_backoff_retry` logger
Explicitly set the name of the `Logger` used by instances of the `Retry`
policies returned by the  `exponential_backoff_retry` factory function.
The name of the logger is now the same as the factory function used to
create those instances, i.e. `exponential_backoff_retry`.

This should improve the readability of the logs and hide the names of
internal non-public classes and functions that make up the
aforementioned `Retry` policy.