feat(utils): make `type_fqn` support callable instances (#43)
Enhance the `sghi.utils.type_fqn` function to accept callables without
the `__qualname__` attribute (non top-level functions and methods). When
given a callable lacking the `__qualname__` attribute, the
representation of the callable (value of `repr(callable)`) is used in
place of `__qualname__` when constructing the fully qualified name of
the callable.